Move marker exceptions into dedicated source set
