+++ testData/markers/org/vineflower/marker/CatchAllException.java
@@ -1,4 +1,4 @@
-package vineflower.markerexception;
+package org.vineflower.marker;
 
 public class CatchAllException extends RuntimeException {
   public CatchAllException(String message) {
