+++ testData/src/java8/pkg/TestTryCatchNoIncrement.java
@@ -1,6 +1,6 @@
 package pkg;
 
-import vineflower.markerexception.UnCatchException;
+import org.vineflower.marker.UnCatchException;
 
 import java.io.File;
 import java.net.URL;
