+++ testData/src/java8/pkg/TestTryFinallyMarkerExceptionsInvalid.java
@@ -1,6 +1,6 @@
 package pkg;
 
-import vineflower.markerexception.CatchAllException;
+import org.vineflower.marker.CatchAllException;
 
 public class TestTryFinallyMarkerExceptionsInvalid {
   public Object testReturn() {
